Implement immersive mode with a mixin starting with android Q
