+++ pubspec.yaml
@@ -68,6 +68,7 @@ dependencies:
   firebase_core: ^2.11.0
   flutter_displaymode: ^0.6.0
   web_socket_channel: ^2.4.0
+  device_info_plus: ^9.0.2
 
 dev_dependencies:
   build_runner: ^2.3.2
